requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
cachetools>=5.3.0
PyMuPDF>=1.23.0
openai>=1.0.0
motor>=3.3.0
